feat(delete vm): added are you sure you want to delete component (#201)
Closes #196
